Revert to Airbnb default for camelcase, add quote-props rule ("consistent")
Fixing all the camelcase violations with properties: always on would have required
breaking the UITK's API. So turn properties: to never, which is Airbnb default,
so remove the rule.

Once you open up the possibility of snake/kebab-case object properties, quoting
becomes an issue again. I don't like the Airbnb config's rule here, it leads to
having to quote some things and not other things. So use this rule with "consistent"
setting, which I've documented in the README.